backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
google-generativeai
